VideoFrames now offset from anchors
XRSession creates and destroys a ongoing series of Anchors, which VideoFrames are expresses relative to.  New anchors are currently created each minute.  These anchors are not exposed.

Rather, session.getVideoFramePose will set in the provided matrix to the current tracker coordinates, by using these anchors to adjust to current time (since the videoframe was captured at some point in the past).